refactor(damage-tracker): make damage tracker use ped memory components as damage booleans
The Damage Tracker used to use health changes to determine whether damage was received. However, with the way Rockstar designed the damage system, the "damageArray", which I'm not sure the real purpose of, is often 0 after the damage history is cleared, allowing for us to check if damage was received by checking that value instead.